w25-project-mod-mocha-main/app/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Django==5.1.6
django-constance==4.3.2
django-extensions==3.2.3
django-rest-framework==0.1.0
djangorestframework==3.15.2
ghp-import==2.1.0
gunicorn==23.0.0
idna==3.10
Jinja2==3.1.5
Markdown==3.7
MarkupSafe==3.0.2
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
PyYAML==6.0.2
pyyaml_env_tag==0.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
watchdog==6.0.0
whitenoise==6.9.0
